Fix data type conversion issue and add file naming functionality

+++ Fall 2023/Code/src/BME680.hpp
@@ -69,7 +69,7 @@ namespace BME
 
     log(bme.temperature);
     log(",");
-    log(float(bme.pressure));
+    log(int(bme.pressure));
     log(",");
     log(bme.humidity);
     log(",");
